rewrites index documents for compatibility with ES7
